+++ server/api/tts.py
@@ -9,7 +9,7 @@ from fastapi.responses import FileResponse
 from server.core.fish_client import FishSpeechClient
 from server.core.voice_manager import VoiceManager
 from server.utils.audio import get_audio_duration
-from shared.models import TTSRequest
+from server.models.tts import TTSRequest
 
 router = APIRouter(prefix="/tts", tags=["tts"])
 fish_client = FishSpeechClient()
