--- shared/__init__.py
@@ -1,1 +1,0 @@
-"""Shared models and utilities for PlomTTS."""
